fix: bug fix on main.py command pair symbol argument

+++ models/price.py
@@ -2,7 +2,7 @@ import uuid
 import datetime
 
 from decouple import config
-from sqlalchemy import Column, create_engine, DateTime, Float, String
+from sqlalchemy import Column, create_engine, DateTime, DECIMAL, String
 from sqlalchemy.dialects.postgresql import UUID
 from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.orm import sessionmaker
@@ -17,9 +17,9 @@ class Price(base):
 
     uuid = Column(UUID(as_uuid=True), primary_key=True, default=uuid.uuid4, unique=True)
     pair = Column(String, nullable=False)
-    curr = Column(Float, nullable=False)
-    lowest = Column(Float, nullable=False)
-    highest = Column(Float, nullable=False)
+    curr = Column(DECIMAL, nullable=False)
+    lowest = Column(DECIMAL, nullable=False)
+    highest = Column(DECIMAL, nullable=False)
     created = Column(DateTime, default=datetime.datetime.utcnow)
